# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Maplewood High School Scheduling API
  description: API for managing school schedules, students, teachers, and classrooms
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
paths:
...
  /students:
    get:
      tags:
        - students
      summary: List students
      operationId: getStudents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentPageResponse'
  /students/{studentId}/schedule/{year}/{semester}:
    get:
      tags:
        - students
      summary: Get student schedule
      operationId: getStudentSchedule
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: semester
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/SemesterSeason'
      responses:
        '200':
          description: Student schedule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentScheduleResponse'
        '404':
          description: Student or schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /students/{studentId}/enrollable/{year}/{semester}:
    get:
      tags:
        - students
      summary: Get available course sections for student
      description: Returns courses that student is eligible to enroll
      operationId: getEligibleCourses
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: semester
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/SemesterSeason'
      responses:
        '200':
          description: Available courses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EligiblePageResponse'
        '404':
          description: Student or semester not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /students/{studentId}/enrollment/{sectionId}:
    post:
      tags:
        - students
      summary: Enroll student in a course section
      operationId: enrollStudent
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student enrolled successfully
        '400':
          description: Enrollment failed (time conflict, no spots, missing prerequisites, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student or section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    SemesterSeason:
      type: string
      enum:
        - FALL
        - SPRING
    Weekday:
      type: string
      enum:
        - MONDAY
        - TUESDAY
        - WEDNESDAY
        - THURSDAY
        - FRIDAY
...
    UnavailabilityReason:
      type: string
      enum:
        - NO_SPOTS
        - TIME_CONFLICT
        - ALREADY_ENROLLED
        - PREREQUISITE_NOT_MET
        - GRADE_LEVEL_REQUIREMENT
        - MAX_ENROLLMENTS_REACHED
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - path
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    PageResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ScheduleDurationResponse:
      type: object
      required:
        - weekday
        - start
        - end
      properties:
        weekday:
          $ref: '#/components/schemas/Weekday'
        start:
          type: integer
        end:
          type: integer
...
    TimeSlotResponse:
      type: object
      required:
        - weekday
        - start
        - end
        - courseCode
        - courseName
        - section
      properties:
        weekday:
          $ref: '#/components/schemas/Weekday'
        start:
          type: integer
        end:
          type: integer
        courseCode:
          type: string
        courseName:
          type: string
        section:
          type: integer
        classroom:
          type: string
        classroomId:
          type: integer
        teacher:
          type: string
        teacherId:
          type: integer
        filledSpots:
          type: integer
        availableSpots:
          type: integer
    StudentResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - fullName
        - email
        - gradeLevel
        - enrollmentYear
        - expectedGraduationYear
        - status
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        email:
          type: string
        gradeLevel:
          type: integer
        enrollmentYear:
          type: integer
        expectedGraduationYear:
          type: integer
        status:
          type: string
    StudentScheduleResponse:
      type: object
      required:
        - studentId
        - studentName
        - timeSlots
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        timeSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlotResponse'
...
    StudentPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/StudentResponse'
...
    EligiblePageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/AvailableCourseSectionResponse'
    AvailableCourseSectionResponse:
      type: object
      required:
        - sectionId
        - courseCode
        - courseName
        - section
        - teacher
        - classroom
        - schedule
        - availableSpots
        - filledSpots
        - available
      properties:
        sectionId:
          type: integer
        courseCode:
          type: string
        courseName:
          type: string
        courseDescription:
          type: string
        credits:
          type: number
        section:
          type: integer
        teacher:
          type: string
        classroom:
          type: string
        schedule:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDurationResponse'
        availableSpots:
          type: integer
        filledSpots:
          type: integer
        available:
          type: boolean
        unavailableReason:
          $ref: '#/components/schemas/UnavailabilityReason'